Added the models folder.
The models folder contains the model experiments.
Each model is contained in its own folder and should
include a file called model.py. A class called Model must
be defined. It should extend the class ModelBase. Each
model should be given an intelligible name. This name
is the name of the folder.

As of now, two models are defined: basic and derivative.
More information can be found in their model.py file.