Support for resizing the canvas.
TomoeCanvas used the window coordinates to append points. Points were
converted to a 1000*1000 basis when get_writing was called. This made
the resized writing incorrect due to precision loss.

Now the canvas uses the 1000*1000 basis internally and points are
converted to this basis when they are appended. Points are converted
back to the window coordinates when they are drawn.