Implementation of a replacement for TomoeChar in pure python and unit tests.
This is one step further toward removing the dependency on Tomoe.
This should fix some issues that were in Tomoe and make the package easy
to use in Windows.

The replacement should be API compatible with Tomoe. In addition to its
coordinates, a point can now define pressure, xtilt, ytilt and
timestamp. Stroke can now have a duration.