Added basic Writing renderers.
A writing renderer takes a writing object and renders it.
Currently 4 renderers are supported:
- Cairo renderer
- PNG renderer
- SVG renderer
- PDF renderer

The last three are all subclasses of the Cairo renderer.

Those renderers will be useful to build the handwriting database website. This will alow to display
a visual representation of people's handwriting contributions.

Ideally, the Canvas in libtegaki-gtk should reuse the Cairo renderer when it redraws itself
completely. But this requires to use Cairo as drawing backend for the Canvas, which is not
the case yet.

Other ideas of renderers include stroke by stroke rendering, visual writing diff
(comparison of two writings) and GIF stroke animation.