More vocabulary.
If Chinese/Japanese people can learn English and non-native speakers
can learn Chinese/Japanese through the process of contributing their handwriting,
this will make contributing handwriting samples much more attractive. This is
why vocabulary lists are important for the tegaki-db project.

For English, it would be possible to build vocabulary lists based on the
EDICT project. It would be great also to include vocabulary lists found
on the internet if their license if free.