[tegaki-db] Removed the vocab/ directory.
At first, I was thinking of letting people add their vocabulary lists so they can be asked to write samples for words from these but
it may be preferrable to leave this role to external games, which can use our web API to submit samples. So I remove the vocab/ directory for now.